+++ conanfile.py
@@ -53,12 +53,12 @@ class MdtApplicationConan(ConanFile):
       # As workaround, try fix a known version that we can build
       # Take a Qt version that we have in our Docker images
       # Hmm, now try to use package from conan-center (20.04.2022)
-      self.requires("qt/5.15.6")
+      self.requires("qt/5.15.16")
 
   def build_requirements(self):
     if self._requires_catch():
-      self.test_requires("catch2/2.13.9")
-    self.test_requires("mdtcmakemodules/0.20.0@scandyna/testing")
+      self.test_requires("catch2/2.13.10")
+    self.test_requires("mdtcmakemodules/0.20.2@scandyna/testing")
 
   def generate(self):
     tc = CMakeToolchain(self)