+++ packaging/conan/mdtguiapplicationfornonqtusage/conanfile.py
@@ -11,6 +11,7 @@ class MdtGuiApplicationForNonQtUsageConan(ConanFile):
   settings = "os", "compiler", "build_type", "arch"
   options = {"shared": [True, False]}
   default_options = {"shared": True}
+  package_type = "library"
   generators = "CMakeDeps", "VirtualBuildEnv"
 
   # See: https://docs.conan.io/en/latest/reference/conanfile/attributes.html#short-paths
@@ -32,12 +33,12 @@ class MdtGuiApplicationForNonQtUsageConan(ConanFile):
     self.output.info( "%s: version is %s" % (self.name, self.version) )
 
   def requirements(self):
-    self.requires("mdtcommandlinearguments/%s@scandyna/testing" % (self.version) )
-    self.requires("mdtcmakeconfig/0.1.0@scandyna/testing")
-    self.requires("qt/5.15.6")
+    self.requires("mdtcommandlinearguments/%s@scandyna/testing" % (self.version), transitive_headers=True)
+    self.requires("mdtcmakeconfig/0.2.3@scandyna/testing")
+    self.requires("qt/5.15.16", transitive_headers=True, transitive_libs=True)
 
   def build_requirements(self):
-    self.test_requires("mdtcmakemodules/0.20.0@scandyna/testing")
+    self.test_requires("mdtcmakemodules/0.20.2@scandyna/testing")
 
   def export_sources(self):
     source_root = os.path.join(self.recipe_folder, "../../../")
@@ -77,6 +78,7 @@ class MdtGuiApplicationForNonQtUsageConan(ConanFile):
     self.cpp_info.set_property("cmake_file_name", "Mdt0GuiApplicationForNonQtUsage")
     self.cpp_info.set_property("cmake_target_name", "Mdt0::GuiApplicationForNonQtUsage")
     # We have to specify the components of Qt to use, otherwise we will depend on all
+    # TODO: check what the status is with Conan 2 about that
     # See also https://gitlab.com/scandyna/mdtapplication/-/issues/4
-    self.cpp_info.requires = ["mdtcommandlinearguments::mdtcommandlinearguments", "qt::qtGui"]
+    self.cpp_info.requires = ["mdtcmakeconfig::mdtcmakeconfig", "mdtcommandlinearguments::mdtcommandlinearguments", "qt::qtGui"]
     self.cpp_info.libs = ["Mdt0GuiApplicationForNonQtUsage"]