+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 
-# [0.5.3] - 2024-12-XX
+# [0.5.3] - 2025-11-16
 
 ## Fixed
 
 - Conan packages: define Qt components in `package_info()` (https://gitlab.com/scandyna/mdtapplication/-/issues/4)
-- Use `mdtcmakemodules` v0.20.0 (https://gitlab.com/scandyna/mdtapplication/-/issues/7)
+
+## Changed
+
+- Update doc to Conan 2 command line (https://gitlab.com/scandyna/mdtapplication/-/issues/7)
+- Update dependencies to use Conan 2 pre-built package (https://gitlab.com/scandyna/mdtapplication/-/issues/7)
 
 
 # [0.5.2] - 2024-11-11